Allow setting game directory from CLI; add -v flag
- Use the directory the binary is in as the default path, so you can use
  "./games/fallout/fallout-ce" without having to cd to that directory.

- Also allow setting the game directory from argv.

- Add -v option for verbose information; this was protected by a _DEBUG
  preprocessor flag, and not having to recompile is helpful.

- Output GNWSystemError() to stderr as well. For whatever reason I don't
  get any popup window (not sure why), so the game would simply exit
  without any information. It doesn't hurt to print it anyway.

- Add -h flag to document all of this.

Note: I only tested this on my Linux system, as that's the only machine
I have access to. In particular, I'm not entirely sure if my usage of
PathRemoveFileSpecA() is correct, but I have no way to test it.

Also clarify which values are read from f1_res.ini, as per my reading of
src/plib/gnw/svga.cc (Fixes #84).