Make arrow keys not scroll so slow you want to self-harm
All keyboard and mouse button work via standard text editor semantics:
keeping a button pressed will wait for a while, after which it will
repeat at fixed rate.

That's fine for a text editor or whatnot, but it's painful when you want
to scroll with the arrow keys. I hated this back when I first played the
game, and I'm hating it now. I don't think anyone will miss the old
"functionality".

We can't outright remove the delay for all keys, as e.g. pressing "i"
will repeat so fast that you can never open the inventory. Come to think
of it, I think we can remove all of this functionality, at least for the
keyboard as I don't *think* it's ever used (it is used for mouse
buttons, although that too can be simplified greatly) – but all of that
is a much larger refactor.